+++ src/completed.js
@@ -19,6 +19,7 @@ export default [
     '-webkit-border-before-style',
     '-webkit-border-before-width',
     '-webkit-mask-attachment',
+    '-webkit-mask-composite',
     '-webkit-mask-repeat',
     '-webkit-mask-repeat-x',
     '-webkit-mask-repeat-y',